+++ CHANGES.rst
@@ -29,6 +29,8 @@ Unreleased
   permitted maximum lengths.
 - Fixed bug where :meth:`.Reddit.request` could not handle instances of ``BadRequest``\s
   when the JSON data contained only the keys "reason" and "message".
+- Fixed bug where :meth:`.Reddit.request` could not handle instances of ``BadRequest``\s
+  when the response did not contain valid JSON data.
 
 7.2.0 (2021/02/25)
 ------------------